test: separate cmap test runner from tests
